Update 2023.09.14.01
Created a struct to hold the background level layer data with a for loop for the properties sharing the same values. Then created a updateLevelData function using that struct to reset the background during animation to imitate a infinite looping background. Then lastly created a float function to set each background layer speed. All functions set to be called in the main game loop