Update 2023.10.07.01
Added, rendered, and updated the enemy sprite to replace the power crystals on collision to end the game. End created two separate collision variables to determine current and future conditionals. Sprite had to be inverted, so had to update the updateAnimationData method to accept a bool to determine whether inversion is truthy then update the animation data depending on the values. Additional refactoring, code cleaning, and optimizing was also done